Add files via upload
Objectives:
The purpose of this activity is to learn the fundamentals of network monitoring and packet sniffing with the help of Wireshark, mainly, and some other tools. We will use Wireshark to explore application layer protocols such as HTTP and DNS.
Question 1: HTTP
Select a website of your choice; must be different than the Wireshark labs above. Now follow the approach suggested in the “Getting Started” Lab above to capture the HTTP GET and HTTP OK messages (one GET and its corresponding OK). Add the following items to your submission:
#1-A: Add the Wireshark screenshot and highlight the two messages in it. [2 Marks]
#1-B: When was this file/object last modified? To support your answer, provide evidence from the captured packets. [2 Marks]
#1-C: Add the answers to the Q12 and Q15 in the HTTP lab. To support your answer, add the screen shot and highlight the field from where you figured out the answer. [2 Marks]
Question 2: DNS
For this question, follow the instructions completely in the “DNS” Lab. Now, add the following items to your submission:
Add your answers to the “DNS” Lab questions 4, 6 and 8 in the submission document. Add the screenshots to support your answers. [1+2+2 Marks]
Question 3: HTTP and DNS
#3-A: Consider the HTTP Request message format (Fig. 2.8 in the book). The header portion has some mandatory fields and then several optional fields; in the last, it has body. Create a Java class, only the data part without any methods, that can be used to store the content of an HTTP Request message. For the optional fields assume that there is no exhaustive list of all possible header fields. [10 Marks]
#3-B: Consider the HTTP Response message format (Fig. 2.9 in the book). The header portion has some mandatory fields and then several optional fields; in the last, it has body. Create a Java class, only the data part without any methods, that can be used to store the content of an HTTP Request message. For the optional fields assume that there is no exhaustive list of all possible header fields. [10 Marks]
#3-C: Consider the DNS message format (Fig. 2.21 in the book). There are some optional fields and some required fields. Create a Java class, only the data part without any methods, that can be used to store the content of a DNS message. [10 Marks]