Add files via upload
Objectives:
The purpose of this activity is to learn the fundamentals of network monitoring and packet sniffing with the help of Wireshark, mainly, and some other tools. We will use Wireshark to explore application layer protocols such as HTTP and DNS.
Question 2: DNS
For this question, follow the instructions completely in the “DNS” Lab. Now, add the following items to your submission:
Add your answers to the “DNS” Lab questions 4, 6 and 8 in the submission document. Add the screenshots to support your answers. [1+2+2 Marks]